Extract NetrunnerHandler to a lib directory
Also added testing.

This necessitated moving to Python3, which triggered some other
changes (e.g. no longer needing the `types` module).

I *think* Python3 was also responsible for `types.MethodType(self, instance, owner)`
no longer working as it once did. But this implementation is how the
StackOverflow answer suggested, anyway, so it's probably better :)

Note that I'm pushing this now just to be able to guarantee that the integration
works, but I'll be following up with a commit intended to run `pytest` in the CodeBuild
process.